Don't try to fill buffer.
On queuing, push the buffer to the busy list when they are full or if we
don't have any more bytes on input. On dequeuing, don't try to dequeue
the requested length but only what is available in the busy list.

On dequeuing, the buffer is left with some garbage at the end. This
doesn't seem to harm the decoding by VLC but this may be a problem.